Refactore app, Provide Retrofit instead of FAN, Migrate from Rx2 to Rx3, implement theMovieDB signUp/Login with local DB prefs and new error handling, remember acc, remove Observables from login screen and more

+++ app/src/main/java/com/android/template/data/models/api/response/LoginResponse.kt
@@ -1,3 +1,3 @@
 package com.android.template.data.models.api.response
 
-data class LoginResponse(val accessToken: String, val expiresIn: Int, val refreshToken: String)
+data class LoginResponse(val accessToken: String, val expiresIn: Long, val refreshToken: String)
